backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
langchain==0.1.13
langchain-groq==0.1.3
langchain-huggingface==0.0.3
sentence-transformers==2.7.0
pydantic==2.6.4
chromadb==0.4.24
pypdf==4.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
